+++ NEWS.md
@@ -1,3 +1,10 @@
+# flandersqmd 0.0.3
+
+* Create Docker for rendering the examples.
+* Fix DOI in colophon only report.
+* Add DOI.
+* Update instructions in README.
+
 # flandersqmd 0.0.2
 
 * `create_report()` doesn't open an empty project (#2).